Extract shared review-gate CI repair path
