+++ packages/execution-engine/src/index.ts
@@ -56,6 +56,7 @@ export * from './worker-types.js';
 export * from './worker-lock.js';
 export * from './builtin-workers.js';
 export * from './auto-fix-recovery.js';
+export * from './review-gate-ci-repair.js';
 export * from './workers/pr-status-worker.js';
 export * from './workers/ci-failure-worker.js';
 export * from './workers/auto-approve-worker.js';
